[bugfix] fn: Align built-in function signatures with W3C XQuery 4.0 PR197
Audit and fix FunctionSignature definitions against the QT4CG XQuery 4.0
spec so partial-application instance-of tests succeed. The QT4 keyword
test catalog (misc-BuiltInKeywords) checks that
fn:foo(arg := ?) instance of function(...) as ... matches the
signature declared by W3C; before this commit eXist's signatures
diverged in parameter names, cardinalities, and return types.

Common patterns fixed:
- collation parameter widened from xs:string to xs:string?
  (fn:contains, fn:ends-with, fn:starts-with, fn:compare,
  fn:distinct-values, fn:index-of, fn:substring-before,
  fn:substring-after, fn:collation-key, fn:string-join's separator,
  array:index-of)
- options map parameter widened to map(*)?
  (fn:doc, fn:doc-available, fn:csv-to-arrays, fn:parse-csv,
  fn:csv-to-xml, fn:csv-doc, fn:parse-xml, fn:parse-xml-fragment,
  fn:path)
- length-style optional trailing parameter widened to ?
  (fn:substring's length, fn:subsequence's length,
  array:subarray's length, array:build's action)
- start type widened to xs:numeric for fn:subsequence
- base param of fn:resolve-uri is xs:string?
- fn:seconds value is xs:decimal?
- fn:unix-dateTime value is xs:nonNegativeInteger?, returns dateTimeStamp
- typed function/array/record/map returns:
  fn:op returns fn(item()*, item()*) as item()*
  fn:invisible-xml returns fn(xs:string) as item()
  fn:analyze-string returns element(fn:analyze-string-result)
  fn:element-to-map returns map(xs:string, item()?)?
  fn:function-annotations returns map(xs:QName, xs:anyAtomicType*)*
  fn:csv-to-arrays returns array(xs:string)*
  fn:divide-decimals returns record(quotient/remainder)
  fn:in-scope-namespaces returns map(xs:NCName, xs:anyURI)
  fn:transitive-closure returns node()*
  array:members returns record(value as item()*)*
  fn:transform returns map(*)
  fn:collation accepts map(*) and returns xs:string

Type-checker support added:
- SequenceType.isSubtypeOf now handles a choice (union) type on the
  sub side: every alternative must be a subtype of the supertype.
  This unblocks date/time accessors (fn:year-from-dateTime etc.) and
  fn:char where the spec types are union types but eXist uses a
  single broader primary type.
- Bare map(*) is treated as map(xs:anyAtomicType, item()*) in subtype
  checks (was xs:string, which contradicts the W3C spec); records flow
  through with an xs:string key fallback because record keys are
  always strings.
- Bare array(*) is now treated as array(item()*); array(*) is a
  subtype of array(item()*).

Implementation tweaks accompanying the signature changes:
- CollatingFunction.getCollator handles an empty xs:string? collation
  argument by returning the default collator.
- FunSubstring, FunSubSequence, FunAnalyzeString, ArrayFunction handle
  an empty optional-length argument by behaving as the no-length form.
- ArrayBuild handles an empty action argument by returning the input
  unchanged (identity-like).
- FunResolveURI handles an empty base-URI argument by falling back to
  the static base URI.

XQTS misc-BuiltInKeywords pass rate: 72.9% -> 89.5% (76 -> 31 fails).
Remaining failures are net-new XQ4 functions (j-tree, j-key, j-value,
j-position, system-properties, build-dateTime, etc.), unrecognised
record types (fn:dateTime-record, fn:parsed-csv-structure-record,
etc.), parser support for record/document-node wildcards in instance-of
expressions, and a fn:matches partial-application internal error —
all out of scope for this signature audit.

Other affected XQTS sets (no regressions, only improvements):
fn-collation-key 10 -> 7, fn-string-join 2 -> 0, fn-resolve-uri 3 -> 1,
fn-substring 1 -> 0, fn-substring-before 3 -> 1, fn-substring-after
2 -> 0, misc-Subtyping 26 -> 25.

+++ exist-core/src/main/java/org/exist/xquery/functions/fn/FnSeconds.java
@@ -36,7 +36,7 @@ public class FnSeconds extends BasicFunction {
             new QName("seconds", Function.BUILTIN_FUNCTION_NS),
             "Returns a dayTimeDuration representing the given number of seconds.",
             new SequenceType[]{
-                    new FunctionParameterSequenceType("value", Type.DOUBLE, Cardinality.ZERO_OR_ONE, "The number of seconds")
+                    new FunctionParameterSequenceType("value", Type.DECIMAL, Cardinality.ZERO_OR_ONE, "The number of seconds")
             },
             new FunctionReturnSequenceType(Type.DAY_TIME_DURATION, Cardinality.ZERO_OR_ONE, "The duration"));
